chore: move registry utilities into their own file and simplify version checks
