--- packages/mcp-provider-api/src/constants.ts
@@ -1,5 +1,0 @@
-import pkg from '../package.json' with { type: 'json' };
-
-const packageJson: {version: string} = pkg as {version: string};
-
-export const MCP_PROVIDER_API_VERSION: string = packageJson.version;
