+++ packages/mcp-provider-api/src/index.ts
@@ -1,14 +1,12 @@
 export {
-    MCP_PROVIDER_API_VERSION
-} from "./constants.js"
-
-export {
     McpPrompt,
     type McpPromptConfig
 } from "./prompts.js"
 
 export {
-    McpProvider
+    MCP_PROVIDER_API_VERSION,
+    McpProvider,
+    type Versioned
 } from "./provider.js"
 
 export {
@@ -18,9 +16,13 @@ export {
 
 export {
     McpTool,
-    type McpToolConfig,
-    Toolset
+    type McpToolConfig
 } from "./tools.js"
+
+export {
+    Toolset,
+    TOOLSETS
+} from "./toolset.js"
 
 export {
     type Services,
