+++ packages/mcp-provider-api/src/tools.ts
@@ -1,7 +1,7 @@
 import { RequestHandlerExtra } from "@modelcontextprotocol/sdk/shared/protocol.js";
 import { CallToolResult, ServerNotification, ServerRequest, ToolAnnotations } from "@modelcontextprotocol/sdk/types.js";
 import { z } from "zod"
-import { MCP_PROVIDER_API_VERSION } from "./constants.js";
+import { Toolset } from "./toolset.js";
 
 export abstract class McpTool<InputArgsShape extends z.ZodRawShape = z.ZodRawShape, OutputArgsShape extends z.ZodRawShape = z.ZodRawShape> {
     abstract getToolsets(): Toolset[]
@@ -14,14 +14,6 @@ export abstract class McpTool<InputArgsShape extends z.ZodRawShape = z.ZodRawSha
         ? [args: z.objectOutputType<InputArgsShape, z.ZodTypeAny>, extra: RequestHandlerExtra<ServerRequest, ServerNotification>]
         : [extra: RequestHandlerExtra<ServerRequest, ServerNotification>]
     ): CallToolResult | Promise<CallToolResult>;
-
-    /**
-     * This method allows the server to check that this instance is compatible and is able to be registered.
-     * Subclasses should not override this method.
-     */
-    public getVersion(): string {
-        return MCP_PROVIDER_API_VERSION;
-    }
 }
 
 export type McpToolConfig<InputArgsShape extends z.ZodRawShape = z.ZodRawShape, OutputArgsShape extends z.ZodRawShape = z.ZodRawShape> = {
@@ -30,15 +22,4 @@ export type McpToolConfig<InputArgsShape extends z.ZodRawShape = z.ZodRawShape, 
     inputSchema?: InputArgsShape;
     outputSchema?: OutputArgsShape;
     annotations?: ToolAnnotations;
-}
-
-// Toolset that a tool should live under
-export enum Toolset {
-    CORE = 'core',
-    DATA = 'data',
-    ORGS = 'orgs',
-    METADATA = 'metadata',
-    TESTING = 'testing',
-    USERS = 'users',
-    EXPERIMENTAL = 'experimental'
 }
